+++ modules/processings/fixLabelPositionCRS.py
@@ -103,4 +103,9 @@ class FixLabelPositionCRS(QgsProcessingAlgorithm):
         return "auxiliar"
 
     def shortHelpString(self):
-        return self.tr("O algoritmo ...")
+        return self.tr("""
+        Este algoritmo corrige a posição de rótulos (label_x e label_y) em camadas vetoriais quando há uma mudança no Sistema de Referência de Coordenadas (CRS).
+        Ele permite selecionar várias camadas de entrada e especificar o CRS antigo para reprojetar as coordenadas dos rótulos para o novo CRS da camada.
+        O algoritmo atualiza os valores dos atributos label_x e label_y com as novas coordenadas, mantendo a precisão da localização dos rótulos após a mudança de CRS.
+        """)
+