+++ modules/processings/placePowerPlantSymbol.py
@@ -169,4 +169,13 @@ class PlacePowerPlantSymbol(QgsProcessingAlgorithm):
         return "edicao"
 
     def shortHelpString(self):
-        return self.tr("O algoritmo ...")
+        return self.tr("""
+        Este algoritmo insere símbolos de elementos de energia em polígonos específicos de infraestrutura energética, conforme seu tamanho e escala de edição.
+                       
+        O algoritmo permite ocultar feições que não possuem tamanho suficiente para o símbolo.
+                       
+        A inserção dos símbolos é realizada no centroide ou ponto interno da feição, ajustando o tamanho do símbolo de acordo com a escala selecionada.
+                       
+        O algoritmo pode ser aplicado em todas as feições ou apenas nas feições selecionadas.
+        """)
+