+++ modules/processings/placeVegetationSymbol.py
@@ -146,4 +146,13 @@ class PlaceVegetationSymbol(QgsProcessingAlgorithm):
         return "edicao"
 
     def shortHelpString(self):
-        return self.tr("O algoritmo ...")
+        return self.tr("""
+        Este algoritmo insere símbolos de vegetação em polígonos de áreas vegetativas, ajustando o tamanho do símbolo de acordo com a escala de edição selecionada.
+                       
+        Os símbolos são posicionados no centroide ou no ponto interno da feição, dependendo da geometria.
+                       
+        O algoritmo identifica o tipo de vegetação através de um mapeamento predefinido e insere o nome correspondente como texto de edição.
+                       
+        Pode ser aplicado em todas as feições ou apenas nas feições selecionadas.
+        """)
+