+++ CHANGELOG.md
@@ -1,7 +1,6 @@
 # CHANGELOG
 
-## 1.2.40
-
+## 1.2.41
 
 Novas Funcionalidades
 
@@ -25,8 +24,9 @@ Melhorias:
 - A ferramenta de pular linhas agora retorna ao estado original caso não tenha como inserir mais quebras;
 - Melhoria no cálculo de posicionamento de rótulo no diagrama de localização do produto;
 - Adicionada generalização automática de drenagens no diagrama de elevação;
-- Melhorada a documentação dos seguintes processings: brdigeAndManholeRotation, bridgeAndManholeWidth, createCustomPolygons, damWidth, loadMasks, placePointOfChange;
+- Melhoria na documentação dos seguintes processings: bridgeAndManholeRotation, bridgeAndManholeWidth, createCustomPolygons, damWidth, loadMasks, placePointOfChange;
 - Melhoria no changeAttributeTopo: há previsão da abreviação dos textos de edição nas classes constr_edificacao_p, constr_edificacao_a e elemnat_trecho_drenagem_l.
+- Melhoria na documentação dos seguintes processings: fixLabelPostionsCRS, placeBorderOfHydrographicElement, placeBuildingSymbol, placeMineralExtractionSymbol, placepowerPlantSymbol, placeVegetationSymbol, rapidsAndWaterfallRotation, reprojectAttributes, setCurveOrientation
 
 Correção de bug: